float values

+++ server/Influx.py
@@ -221,22 +221,22 @@ class Influx(threading.Thread):
 
                 if keys[2] == "SDS_P1":
                     json_body[0]['measurement'] = "outside-pm10"
-                    json_body[0]['fields']['value'] = int(v)
+                    json_body[0]['fields']['value'] = float(v)
                     self._influx.write_points(json_body)
 
                 if keys[2] == "SDS_P2":
                     json_body[0]['measurement'] = "outside-pm25"
-                    json_body[0]['fields']['value'] = int(v)
+                    json_body[0]['fields']['value'] = float(v)
                     self._influx.write_points(json_body)
 
                 if keys[2] == "temperature":
                     json_body[0]['measurement'] = "outside-temperature"
-                    json_body[0]['fields']['value'] = int(v)
+                    json_body[0]['fields']['value'] = float(v)
                     self._influx.write_points(json_body)
 
                 if keys[2] == "humidity":
                     json_body[0]['measurement'] = "outside-humidity"
-                    json_body[0]['fields']['value'] = int(v)
+                    json_body[0]['fields']['value'] = float(v)
                     self._influx.write_points(json_body)
 
                 if keys[2] == "BMP_pressure":
@@ -248,7 +248,7 @@ class Influx(threading.Thread):
 
                 if keys[1] == "speed":
                     json_body[0]['measurement'] = "bike-speed"
-                    json_body[0]['fields']['value'] = int(v)
+                    json_body[0]['fields']['value'] = float(v)
                     self._influx.write_points(json_body)
